feat!: added `timestamp` field to API response (#15)
BREAKING CHANGE: `message` field has been renamed to `msg`

Let's pretend that in order to add the `timestamp` field, we had to also
change the `message` field to `msg`. The addition of a new field to an
API reponse body would be considered a feature, but changing an existing
key would be a breaking change. This changeset represents both what a
customer would see when there's a new feature, but also when there's a
breaking change.